addresses simba-git comments on PR
